# server/docs/openapi.yaml
openapi: 3.0.0
info:
  title: API do Sistema Financeiro
  description: |
    API completa para gerenciamento de finanças pessoais e empresariais.
    
    ## Funcionalidades Principais
    
    ### Gestão Financeira
    - **Contas**: Gerenciamento de contas bancárias e carteiras
    - **Transações**: Registro de receitas e despesas
    - **Categorias**: Organização de transações por categorias
    - **Relatórios**: Análises e relatórios financeiros
    
    ### Gestão de Clientes e Fornecedores
    - **Clientes**: Cadastro e gestão de clientes
    - **Fornecedores**: Controle de fornecedores
    - **Recebíveis**: Controle de contas a receber
    - **Pagáveis**: Controle de contas a pagar
    
    ### Investimentos e Financiamentos
    - **Investimentos**: Gestão de portfólio de investimentos
    - **Metas de Investimento**: Definição e acompanhamento de metas
    - **Financiamentos**: Controle de empréstimos e financiamentos
    - **Pagamentos**: Gestão de parcelas e pagamentos
    
    ### Recursos Avançados
    - **Autenticação JWT**: Sistema seguro de autenticação
    - **Notificações**: Sistema de alertas e lembretes
    - **Jobs**: Processamento em background
    - **Validações**: Validação robusta de dados
    
    ## Status dos Testes
    - **13/17 suítes de integração** funcionando 100%
    - **142/215 testes** passando (66%)
    - **Execução sequencial** implementada para evitar conflitos
    - **Cobertura de código** configurada
    
    ## Documentação Relacionada
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint de login
...
security:
  - bearerAuth: []
...
paths:
...
  /api/financings:
    get:
      tags:
        - Financings
      summary: Listar financiamentos
      description: Lista todos os financiamentos do usuário com filtros e paginação
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Número da página
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Limite de itens por página
        - name: financing_type
          in: query
          schema:
            type: string
            enum: [hipoteca, emprestimo_pessoal, veiculo, outros]
          description: Filtrar por tipo de financiamento
        - name: creditor_id
          in: query
          schema:
            type: integer
          description: Filtrar por credor
        - name: status
          in: query
          schema:
            type: string
            enum: [ativo, quitado, inadimplente]
          description: Filtrar por status
        - name: amortization_method
          in: query
          schema:
            type: string
            enum: [SAC, Price]
          description: Filtrar por método de amortização
        - name: start_date_from
          in: query
          schema:
            type: string
            format: date
          description: Data de início a partir de
        - name: start_date_to
          in: query
          schema:
            type: string
            format: date
          description: Data de início até
      responses:
        '200':
          description: Lista de financiamentos
    post:
      tags:
        - Financings
      summary: Criar financiamento
      description: Cria um novo financiamento
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                creditor_id:
                  type: integer
                  description: ID do credor
                financing_type:
                  type: string
                  enum: [hipoteca, emprestimo_pessoal, veiculo, outros]
                  description: Tipo de financiamento
                total_amount:
                  type: number
                  minimum: 0.01
                  description: Valor total financiado
                interest_rate:
                  type: number
                  minimum: 0
                  description: Taxa de juros anual (decimal)
                term_months:
                  type: integer
                  minimum: 1
                  description: Prazo em meses
                start_date:
                  type: string
                  format: date
                  description: Data de início do financiamento
                description:
                  type: string
                  description: Descrição do financiamento
                contract_number:
                  type: string
                  description: Número do contrato
                payment_method:
                  type: string
                  enum: [boleto, debito_automatico, cartao, pix]
                  description: Método de pagamento
                observations:
                  type: string
                  description: Observações adicionais
                amortization_method:
                  type: string
                  enum: [SAC, Price]
                  description: Método de amortização
              required:
                - creditor_id
                - financing_type
                - total_amount
                - interest_rate
                - term_months
                - start_date
                - amortization_method
      responses:
        '201':
          description: Financiamento criado com sucesso
  /api/financings/statistics:
    get:
      tags:
        - Financings
      summary: Estatísticas dos financiamentos
      description: Obtém estatísticas dos financiamentos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Estatísticas retornadas
  /api/financings/{id}:
    get:
      tags:
        - Financings
      summary: Obter financiamento
      description: Obtém um financiamento específico
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dados do financiamento
    put:
      tags:
        - Financings
      summary: Atualizar financiamento
      description: Atualiza um financiamento existente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Financiamento atualizado
    delete:
      tags:
        - Financings
      summary: Excluir financiamento
      description: Remove um financiamento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Financiamento removido
  /api/financings/{id}/amortization-table:
    get:
      tags:
        - Financings
      summary: Tabela de amortização
      description: Obtém a tabela de amortização de um financiamento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tabela de amortização
...